make many improvements like prevent xss and sql injection, use sonarQube for code quality and more
