--- src/main/java/com/mohamed/fullTestingDemo/configuration/AsyncConfiguration.java
@@ -1,21 +1,0 @@
-package com.mohamed.fullTestingDemo.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-public class AsyncConfiguration {
-    @Bean
-    public Executor taskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(5);
-        executor.setMaxPoolSize(10);
-        executor.setQueueCapacity(50);
-        executor.setThreadNamePrefix("AsyncExecutor-");
-        executor.initialize();
-        return executor;
-    }
-}
