--- src/main/java/com/mohamed/fullTestingDemo/user/UserRepository.java
@@ -1,11 +1,0 @@
-package com.mohamed.fullTestingDemo.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
