--- src/main/java/com/mohamed/fullTestingDemo/user/dto/request/UpdateUserRequestDto.java
@@ -1,14 +1,0 @@
-package com.mohamed.fullTestingDemo.user.dto.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record UpdateUserRequestDto(
-        @NotBlank
-        String username,
-
-        @NotBlank
-        @Email
-        String email
-){
-}
