--- src/main/java/com/mohamed/fullTestingDemo/user/dto/response/CreateUserResponseDto.java
@@ -1,11 +1,0 @@
-package com.mohamed.fullTestingDemo.user.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record CreateUserResponseDto(
-        Long id,
-        String username,
-        String email
-) {
-}
