--- src/main/java/com/mohamed/fullTestingDemo/user/dto/response/GetUsersResponseDto.java
@@ -1,8 +1,0 @@
-package com.mohamed.fullTestingDemo.user.dto.response;
-
-public record GetUsersResponseDto(
-        Long id,
-        String username,
-        String email
-) {
-}
