--- src/main/java/com/mohamed/fullTestingDemo/user/exceptions/UserAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.mohamed.fullTestingDemo.user.exceptions;
-
-public class UserAlreadyExistsException extends RuntimeException {
-    public UserAlreadyExistsException(String msg){
-        super(msg);
-    }
-}
