--- src/main/java/com/mohamed/fullTestingDemo/user/exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.mohamed.fullTestingDemo.user.exceptions;
-
-public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(String msg){
-        super(msg);
-    }
-}
