+++ src/main/java/com/mohamed/fulltestingdemo/user/UserRepository.java
@@ -1,0 +1,14 @@
+package com.mohamed.fulltestingdemo.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    @Query("SELECT u FROM User u WHERE u.username = :username ")
+    Optional<User> findByUsername(@Param("username") String username);
+}
