+++ src/main/java/com/mohamed/fulltestingdemo/user/dto/response/CreateUserResponseDto.java
@@ -1,0 +1,11 @@
+package com.mohamed.fulltestingdemo.user.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record CreateUserResponseDto(
+        Long id,
+        String username,
+        String email
+) {
+}
