+++ src/main/java/com/mohamed/fulltestingdemo/user/dto/response/GetUsersResponseDto.java
@@ -1,0 +1,8 @@
+package com.mohamed.fulltestingdemo.user.dto.response;
+
+public record GetUsersResponseDto(
+        Long id,
+        String username,
+        String email
+) {
+}
