+++ src/main/java/com/mohamed/fulltestingdemo/user/exceptions/UserAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.mohamed.fulltestingdemo.user.exceptions;
+
+public class UserAlreadyExistsException extends RuntimeException {
+    public UserAlreadyExistsException(String msg){
+        super(msg);
+    }
+}
