+++ src/main/java/com/mohamed/fulltestingdemo/user/exceptions/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.mohamed.fulltestingdemo.user.exceptions;
+
+public class UserNotFoundException extends RuntimeException{
+    public UserNotFoundException(String msg){
+        super(msg);
+    }
+}
