--- src/main/resources/application-dev.properties
@@ -1,10 +1,0 @@
-server.port=8081
-
-## database ##
-spring.datasource.url=jdbc:postgresql://localhost:5432/mydatabase
-spring.datasource.username=myuser
-spring.datasource.password=secret
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.flyway.enabled=false
