+++ src/main/resources/application-test.yml
@@ -1,0 +1,17 @@
+server:
+  port: 7070
+
+spring:
+  datasource:
+    url: jdbc:tc:postgresql:latest://localhost/testdb
+    username: test
+    password: test
+  jpa:
+    database-platform: org.hibernate.dialect.PostgreSQLDialect
+    hibernate:
+      ddl-auto: create-drop
+  flyway:
+    enabled: false
+  test:
+    database:
+      replace: none  # Ensures Testcontainers is used instead of replacing with H2
