--- src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.application.name=fullTestingDemo
-spring.profiles.active=dev
