+++ src/main/resources/application.yml
@@ -1,0 +1,13 @@
+spring:
+  application:
+    name: fullTestingDemo
+  datasource:
+    driver-class-name: org.postgresql.Driver
+  profiles:
+    active: dev
+
+server:
+  compression:
+    enabled: true
+    mime-types: text/html,text/xml,text/plain,text/css,text/javascript,application/javascript,application/json
+    min-response-size: 1024
