--- src/test/java/com/mohamed/fullTestingDemo/TestFullTestingDemoApplication.java
@@ -1,12 +1,0 @@
-package com.mohamed.fullTestingDemo;
-
-import com.mohamed.fullTestingDemo.configuration.TestcontainersConfiguration;
-import org.springframework.boot.SpringApplication;
-
-public class TestFullTestingDemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.from(FullTestingDemoApplication::main).with(TestcontainersConfiguration.class).run(args);
-	}
-
-}
