--- src/test/java/com/mohamed/fullTestingDemo/bdd/steps/CucumberSpringConfiguration.java
@@ -1,9 +1,0 @@
-package com.mohamed.fullTestingDemo.bdd.steps;
-
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@CucumberContextConfiguration
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class CucumberSpringConfiguration {
-}
