+++ src/test/java/com/mohamed/fulltestingdemo/bdd/CucumberTest.java
@@ -1,0 +1,14 @@
+package com.mohamed.fulltestingdemo.bdd;
+
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(
+        features = "src/test/resources/features",
+        glue = "com.mohamed.fullTestingDemo.bdd.steps",
+        plugin = {"pretty", "json:target/cucumber-report.json"}
+)
+public class CucumberTest {
+}
