+++ src/test/java/com/mohamed/fulltestingdemo/bdd/steps/CucumberSpringConfiguration.java
@@ -1,0 +1,9 @@
+package com.mohamed.fulltestingdemo.bdd.steps;
+
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@CucumberContextConfiguration
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class CucumberSpringConfiguration {
+}
