+++ src/test/java/com/mohamed/fulltestingdemo/configuration/TestcontainersConfiguration.java
@@ -1,0 +1,36 @@
+package com.mohamed.fulltestingdemo.configuration;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.utility.DockerImageName;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class TestcontainersConfiguration {
+
+	@Value("${spring.datasource.username}")
+	private String dbUserName;
+
+	@Value("${spring.datasource.password}")
+	private String dbPassword;
+
+	@Bean
+	@ServiceConnection
+	PostgreSQLContainer<?> postgresContainer() {
+		return new PostgreSQLContainer<>(DockerImageName.parse("postgres:latest"))
+				.withDatabaseName("mydatabase")  // ✅ Match docker-compose DB name
+				.withUsername(dbUserName)         // ✅ Match docker-compose user
+				.withPassword(dbPassword);
+	}
+
+	@Bean
+	@ServiceConnection(name = "redis")
+	GenericContainer<?> redisContainer() {
+		return new GenericContainer<>(DockerImageName.parse("redis:latest")).withExposedPorts(6379);
+	}
+
+
+}
